DarkAgesModule/requirements.txt
numpy
scipy
PyYAML
dill
future
